requirements.txt
# Основные зависимости
openai>=1.0.0              # OpenAI API для транскрипции и классификации
python-dotenv>=1.0.0       # Загрузка переменных окружения из .env файла

# Опциональные зависимости
pydub>=0.25.1              # Конвертация аудиофайлов (опционально)
